Self-supervised Video Prediction
Use DSSIM+L1+L2 loss for training. Find the right balance of the gains.

Generate some GIF of the results on test dataset. Similar to (note that this is on soccer but your dataset is different!):
Pay attention that two of the ladder layers have location dependant-convolution paper: http://ais.uni-bonn.de/~hfarazi/papers/LocDep.pdf Code: https://github.com/AIS-Bonn/LocDepVideoPrediction

Use three seed frames and predict the next three frames. The original resolution is 320 × 240, change it so that it is compatible with the model.

After training the video prediction model, add two linear layers on top of the hidden states from Conv-GRU, and classify the actions. You have to forward the full video to the model and then save the Conv-GRU's internal hidden states for that video. Finally, classify the video's action using the save hidden states with an MLP with one hidden layer.

Measure the accuracy of the action classification on the test set.

Try to adjust the hyper-parameters of the model to achieve the best performance possible.